+++ example/bin/main_built_value.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:built_collection/built_collection.dart';
 import 'package:chopper/chopper.dart';
 import 'package:chopper_example/built_value_resource.dart';