+++ chopper/lib/src/request.dart
@@ -27,21 +27,6 @@ class Request {
     this.parts: const [],
   });
 
-  String _getMethod(String method) {
-    switch (method) {
-      case HttpMethod.Delete:
-        return "DELETE";
-      case HttpMethod.Patch:
-        return "PATCH";
-      case HttpMethod.Post:
-        return "POST";
-      case HttpMethod.Put:
-        return "PUT";
-      default:
-        return "GET";
-    }
-  }
-
   Request replace({
     HttpMethod method,
     String url,
@@ -64,98 +49,29 @@ class Request {
         multipart: multipart ?? this.multipart,
       );
 
-  Uri _buildUri() {
-    var uri;
-    if (url.startsWith('http://') || url.startsWith('https://')) {
-      // if the request's url is already a fully qualified URL, we can use
-      // as-is and ignore the baseUrl
-      uri = Uri.parse(url);
-    } else {
-      if (!baseUrl.endsWith('/') && !url.startsWith('/')) {
-        uri = Uri.parse("$baseUrl/$url");
-      } else {
-        uri = Uri.parse("$baseUrl$url");
-      }
-    }
-
-    if (parameters.isNotEmpty) {
-      return uri.replace(
-        queryParameters: parameters.map((k, v) => MapEntry(k, "$v")),
-      );
-    }
-    return uri;
-  }
+  Uri _buildUri() => buildUri(baseUrl, url, parameters);
 
   Map<String, String> _buildHeaders() => Map<String, String>.from(headers);
 
-  Future<http.BaseRequest> _toMultipartRequest(
-    String method,
-    Uri uri,
-    Map<String, String> headers,
-  ) async {
-    final baseRequest = http.MultipartRequest(method, uri);
-    baseRequest.headers.addAll(headers);
-
-    for (final part in parts) {
-      if (part is PartFile) {
-        if (part.value is List<int>) {
-          baseRequest.files.add(
-            http.MultipartFile.fromBytes(part.name, part.value),
-          );
-        } else if (part.value is String) {
-          baseRequest.files.add(
-            await http.MultipartFile.fromPath(part.name, part.value),
-          );
-        }
-      } else {
-        baseRequest.fields[part.name] = part.value?.toString();
-      }
-    }
-    return baseRequest;
-  }
-
-  Future<http.BaseRequest> _toBaseRequest(
-    String method,
-    Uri uri,
-    Map<String, String> headers,
-  ) async {
-    final baseRequest = http.Request(_getMethod(method), uri);
-    baseRequest.headers.addAll(headers);
-
-    if (body != null) {
-      if (body is String) {
-        baseRequest.body = body as String;
-      } else if (body is List) {
-        baseRequest.bodyBytes = (body as List<int>).cast<int>();
-      } else if (body is Map) {
-        baseRequest.bodyFields = (body as Map).cast<String, String>();
-      } else {
-        throw ArgumentError('Invalid request body "${body}".');
-      }
-    }
-    return baseRequest;
-  }
-
-  Future<http.BaseRequest> toHttpRequest() async {
+  Future<http.BaseRequest> toBaseRequest() async {
     final uri = _buildUri();
-    final met = _getMethod(method);
     final heads = _buildHeaders();
 
     if (multipart) {
-      return _toMultipartRequest(
-        met,
+      return toMultipartRequest(
+        parts,
+        method,
         uri,
         heads,
       );
     }
-    return _toBaseRequest(
-      met,
+    return toHttpRequest(
+      body,
+      method,
       uri,
       heads,
     );
   }
-
-  bool get isJson => headers[contentTypeKey] == jsonHeaders;
 }
 
 class HttpMethod {
@@ -185,4 +101,78 @@ class PartValue<T> {
 @immutable
 class PartFile<T> extends PartValue<T> {
   PartFile(String name, T value) : super(name, value);
+}
+
+Uri buildUri(String baseUrl, String url, Map<String, dynamic> parameters) {
+  var uri;
+  if (url.startsWith('http://') || url.startsWith('https://')) {
+    // if the request's url is already a fully qualified URL, we can use
+    // as-is and ignore the baseUrl
+    uri = Uri.parse(url);
+  } else {
+    if (!baseUrl.endsWith('/') && !url.startsWith('/')) {
+      uri = Uri.parse("$baseUrl/$url");
+    } else {
+      uri = Uri.parse("$baseUrl$url");
+    }
+  }
+
+  final params = Map<String, dynamic>.from(parameters);
+  params.removeWhere((k, v) => v == null);
+  if (params.isNotEmpty) {
+    return uri.replace(
+      queryParameters: params.map((k, v) => MapEntry(k, "$v")),
+    );
+  }
+  return uri;
+}
+
+Future<http.Request> toHttpRequest(
+  body,
+  String method,
+  Uri uri,
+  Map<String, String> headers,
+) async {
+  final baseRequest = http.Request(method, uri);
+  baseRequest.headers.addAll(headers);
+
+  if (body != null) {
+    if (body is String) {
+      baseRequest.body = body;
+    } else if (body is List<int>) {
+      baseRequest.bodyBytes = body;
+    } else if (body is Map<String, String>) {
+      baseRequest.bodyFields = body;
+    } else {
+      throw ArgumentError.value('$body', 'body');
+    }
+  }
+  return baseRequest;
+}
+
+Future<http.MultipartRequest> toMultipartRequest(
+  List parts,
+  String method,
+  Uri uri,
+  Map<String, String> headers,
+) async {
+  final baseRequest = http.MultipartRequest(method, uri);
+  baseRequest.headers.addAll(headers);
+
+  for (final part in parts) {
+    if (part is PartFile) {
+      if (part.value is List<int>) {
+        baseRequest.files.add(
+          http.MultipartFile.fromBytes(part.name, part.value),
+        );
+      } else if (part.value is String) {
+        baseRequest.files.add(
+          await http.MultipartFile.fromPath(part.name, part.value),
+        );
+      }
+    } else {
+      baseRequest.fields[part.name] = part.value?.toString();
+    }
+  }
+  return baseRequest;
 }